+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/disconnectedOperationsManagementClient.ts
@@ -2,10 +2,10 @@
 // Licensed under the MIT License.
 
 import type {
-  DisconnectedOperationsMgmtContext,
-  DisconnectedOperationsMgmtClientOptionalParams,
+  DisconnectedOperationsManagementContext,
+  DisconnectedOperationsManagementClientOptionalParams,
 } from "./api/index.js";
-import { createDisconnectedOperationsMgmt } from "./api/index.js";
+import { createDisconnectedOperationsManagement } from "./api/index.js";
 import type { ArtifactsOperations } from "./classic/artifacts/index.js";
 import { _getArtifactsOperations } from "./classic/artifacts/index.js";
 import type { DisconnectedOperationsOperations } from "./classic/disconnectedOperations/index.js";
@@ -17,10 +17,10 @@ import { _getImagesOperations } from "./classic/images/index.js";
 import type { TokenCredential } from "@azure/core-auth";
 import type { Pipeline } from "@azure/core-rest-pipeline";
 
-export { DisconnectedOperationsMgmtClientOptionalParams } from "./api/disconnectedOperationsMgmtContext.js";
+export { DisconnectedOperationsManagementClientOptionalParams } from "./api/disconnectedOperationsManagementContext.js";
 
-export class DisconnectedOperationsMgmtClient {
-  private _client: DisconnectedOperationsMgmtContext;
+export class DisconnectedOperationsManagementClient {
+  private _client: DisconnectedOperationsManagementContext;
   /** The pipeline used by this client to make requests */
   public readonly pipeline: Pipeline;
 
@@ -28,13 +28,13 @@ export class DisconnectedOperationsMgmtClient {
   constructor(
     credential: TokenCredential,
     subscriptionId: string,
-    options: DisconnectedOperationsMgmtClientOptionalParams = {},
+    options: DisconnectedOperationsManagementClientOptionalParams = {},
   ) {
     const prefixFromOptions = options?.userAgentOptions?.userAgentPrefix;
     const userAgentPrefix = prefixFromOptions
       ? `${prefixFromOptions} azsdk-js-client`
       : `azsdk-js-client`;
-    this._client = createDisconnectedOperationsMgmt(credential, subscriptionId, {
+    this._client = createDisconnectedOperationsManagement(credential, subscriptionId, {
       ...options,
       userAgentOptions: { userAgentPrefix },
     });